Move ListWatchUntil to its kin
